Update class.tx_dmmjobcontrol_pi1.php
Implement a new feature: sorting of search results in frontend by adding Get vars like this:
?tx_dmmjobcontrol_pi1[sort]=region&tx_dmmjobcontrol_pi1[sort_order]=asc 

This way we cann offer to sort the search results list by the frontend user. We just added some links in the list template:
        ###REGION_LABEL### <a href="jobs/?tx_dmmjobcontrol_pi1[sort]=region&amp;tx_dmmjobcontrol_pi1[sort_order]=asc" title="sort ascending"><img src="fileadmin/templates/dmmjobcontrol/up.gif"></a><a href="jobs/?tx_dmmjobcontrol_pi1[sort]=region&amp;tx_dmmjobcontrol_pi1[sort_order]=desc" title="sort descending"><img src="fileadmin/templates/dmmjobcontrol/down.gif"></a>


Actually this implementation was made by one of the former extension programmers and we paid for it. They promised to implement it for the public so everybody can benefit from the effort made. So let's try again...